beamer: move get_commit_id from beamer.deploy.util to beamer.util
